requirements.txt
beautifulsoup4==4.10.0
requests==2.27.1
pillow==9.0.1
telethon==1.24.0
python-dotenv==0.19.2
schedule==1.1.0